+++ core-bundle/tests/DependencyInjection/ContaoCoreExtensionTest.php
@@ -315,8 +315,7 @@ class ContaoCoreExtensionTest extends TestCase
         $tags = $definition->getTags();
 
         $this->assertArrayHasKey('doctrine.event_listener', $tags);
-        $this->assertSame('onSchemaIndexDefinition', $tags['doctrine.event_listener'][0]['event']);
-        $this->assertSame('postGenerateSchema', $tags['doctrine.event_listener'][1]['event']);
+        $this->assertSame('postGenerateSchema', $tags['doctrine.event_listener'][0]['event']);
     }
 
     public function testRegistersTheExceptionConverterListener(): void