+++ core-bundle/src/Resources/contao/forms/Form.php
@@ -296,8 +296,6 @@ class Form extends \Hybrid
 			/** @var PageModel $objTarget */
 			$this->Template->action = $objTarget->getFrontendUrl();
 		}
-
-		return $this->Template->parse();
 	}
 
 	/**