+++ calendar-bundle/src/Resources/contao/dca/tl_calendar_events.php
@@ -573,7 +573,7 @@ class tl_calendar_events extends Backend
 		{
 			case 'paste':
 			case 'select':
-				if (!\in_array($id, $root))
+				if (!\in_array(CURRENT_ID, $root)) // check CURRENT_ID here (see #247)
 				{
 					throw new Contao\CoreBundle\Exception\AccessDeniedException('Not enough permissions to access calendar ID ' . $id . '.');
 				}