+++ news-bundle/src/Resources/contao/dca/tl_news.php
@@ -527,7 +527,7 @@ class tl_news extends Backend
 		{
 			case 'paste':
 			case 'select':
-				if (!\in_array($id, $root))
+				if (!\in_array(CURRENT_ID, $root)) // check CURRENT_ID here (see #247)
 				{
 					throw new Contao\CoreBundle\Exception\AccessDeniedException('Not enough permissions to access news archive ID ' . $id . '.');
 				}